Initial draft of OF-Config 1.3 data model:
onf-of-config.yang  -- main module
onf-of-types.yang -- OpenFlow-related type and identity definitions
of-config-1.3.html -- html tree view of the current snapshot

The model has a dependency on the IETF interfaces model (RFC 7223).

This version is still work-in-progress with a several open issues and missing
items, in particular where to model flow-table configuraion, and inclusion of
OF tunnels / logical ports.